refactor/#6 [rename] : 로직 설계 변경으로 인한 chat 코드 수정
