--- app/src/main/java/com/devnuts/ruflu/ui/chat/viewmodel/ChatSubNBViewModel.kt
@@ -1,7 +1,0 @@
-package com.devnuts.ruflu.ui.chat.viewmodel
-
-import androidx.lifecycle.ViewModel
-
-class ChatSubNBViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
