+++ app/src/main/res/layout/fragment_chat.xml
@@ -1,57 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.fragment.HomeFragment"
-    android:background="#ffffff">
+    tools:context=".ui.chat.fragment.ChatFragment">
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/tab_top_layout"
-        android:orientation="horizontal"
-        android:weightSum="1"
-        >
-
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/chat_tabLayout"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            app:tabTextColor="@color/black"
-            android:layout_weight="0.3"
-            >
-
-            <com.google.android.material.tabs.TabItem
-                android:id="@+id/chat_ruflu"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Ruflu"
-                />
-            <com.google.android.material.tabs.TabItem
-                android:id="@+id/chat_NB"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="NearBy"
-                />
-
-        </com.google.android.material.tabs.TabLayout>
-
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:text=""
-            android:textColor="@color/black"
-            android:layout_weight="0.7"
-            />
-    </LinearLayout>
-
-    <androidx.viewpager2.widget.ViewPager2
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/chat_room_recycler"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/chat_viewpager"
-        android:layout_below="@+id/chat_tabLayout"/>
+        android:background="@color/white">
+    </androidx.recyclerview.widget.RecyclerView>
 
-</RelativeLayout>
+    <RelativeLayout
+        android:id="@+id/my_room_container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/white"
+        android:visibility="gone">
+    </RelativeLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
